Implement concurrent_queue in terms of concurrent_deque
