+++ test/CMakeLists.txt
@@ -4,7 +4,6 @@ set(func_sourceFiles
     "func/low_level/test_semaphore.cpp"
     "func/low_level/test_mutexes.cpp"
     "func/data/test_concurrent_dequeue.cpp"
-    "func/data/test_concurrent_queue.cpp"
     "func/detail/test_worker_tasks.cpp"
     "func/test_inline_executor.cpp"
     "func/test_init.cpp"
