+++ test/perf/perf_queue.cpp
@@ -71,8 +71,8 @@ void push(Q& queue, T&& elem) {
     queue.push_back(std::forward<T>(elem));
 }
 
-template <typename T, concore::queue_type QT>
-void push(concore::concurrent_queue<T, QT>& queue, T&& elem) {
+template <typename T>
+void push(concore::concurrent_queue<T>& queue, T&& elem) {
     queue.push(std::forward<T>(elem));
 }
 
@@ -81,8 +81,8 @@ bool try_pop(Q& queue, T& elem) {
     return queue.try_pop_front(elem);
 }
 
-template <typename T, concore::queue_type QT>
-bool try_pop(concore::concurrent_queue<T, QT>& queue, T& elem) {
+template <typename T>
+bool try_pop(concore::concurrent_queue<T>& queue, T& elem) {
     return queue.try_pop(elem);
 }
 
@@ -128,12 +128,8 @@ void BM_push_back_concurrent_dequeue(benchmark::State& state) {
 }
 void BM_push_back_concurrent_queue(benchmark::State& state) {
     CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<test_elem> queue;
-    test_push_back_latency(queue, state);
-}
-void BM_push_back_concurrent_queue_sc(benchmark::State& state) {
-    CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<test_elem, concore::queue_type::multi_prod_single_cons> queue;
+    const int reserved = state.range(1);
+    concore::concurrent_queue<test_elem> queue(reserved);
     test_push_back_latency(queue, state);
 }
 
@@ -176,12 +172,8 @@ void BM_pop_front_concurrent_dequeue(benchmark::State& state) {
 }
 void BM_pop_front_concurrent_queue(benchmark::State& state) {
     CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<test_elem> queue;
-    test_pop_front_latency(queue, state);
-}
-void BM_pop_front_concurrent_queue_sc(benchmark::State& state) {
-    CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<test_elem, concore::queue_type::multi_prod_single_cons> queue;
+    const int reserved = state.range(1);
+    concore::concurrent_queue<test_elem> queue(reserved);
     test_pop_front_latency(queue, state);
 }
 
@@ -258,12 +250,8 @@ void BM_push_pop_par_small_concurrent_dequeue(benchmark::State& state) {
 }
 void BM_push_pop_par_small_concurrent_queue(benchmark::State& state) {
     CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<int> queue;
-    test_push_pop(queue, state);
-}
-void BM_push_pop_par_small_concurrent_queue_sc(benchmark::State& state) {
-    CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<int, concore::queue_type::multi_prod_single_cons> queue;
+    const int reserved = state.range(1);
+    concore::concurrent_queue<int> queue(reserved);
     test_push_pop(queue, state);
 }
 
@@ -280,12 +268,8 @@ void BM_push_pop_par_large_concurrent_dequeue(benchmark::State& state) {
 }
 void BM_push_pop_par_large_concurrent_queue(benchmark::State& state) {
     CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<test_elem> queue;
-    test_push_pop(queue, state);
-}
-void BM_push_pop_par_large_concurrent_queue_sc(benchmark::State& state) {
-    CONCORE_PROFILING_FUNCTION();
-    concore::concurrent_queue<test_elem, concore::queue_type::multi_prod_single_cons> queue;
+    const int reserved = state.range(1);
+    concore::concurrent_queue<test_elem> queue(reserved);
     test_push_pop(queue, state);
 }
 
@@ -301,32 +285,25 @@ void BM_push_pop_par_large_concurrent_queue_sc(benchmark::State& state) {
 BENCHMARK_CASE1(BM_push_back_mtx_queue);
 BENCHMARK_CASE1(BM_push_back_concurrent_dequeue);
 BENCHMARK_CASE1(BM_push_back_concurrent_queue);
-BENCHMARK_CASE1(BM_push_back_concurrent_queue_sc);
 BENCHMARK_CASE2(BM_push_back_mtx_queue);
 BENCHMARK_CASE2(BM_push_back_concurrent_dequeue);
 BENCHMARK_CASE2(BM_push_back_concurrent_queue);
-BENCHMARK_CASE2(BM_push_back_concurrent_queue_sc);
 BENCHMARK_CASE3(BM_push_back_mtx_queue);
 BENCHMARK_CASE3(BM_push_back_concurrent_dequeue);
 BENCHMARK_CASE3(BM_push_back_concurrent_queue);
-BENCHMARK_CASE3(BM_push_back_concurrent_queue_sc);
 BENCHMARK_CASE4(BM_push_back_mtx_queue);
 BENCHMARK_CASE4(BM_push_back_concurrent_dequeue);
 BENCHMARK_CASE4(BM_push_back_concurrent_queue);
-BENCHMARK_CASE4(BM_push_back_concurrent_queue_sc);
 
 BENCHMARK_CASE1(BM_pop_front_mtx_queue);
 BENCHMARK_CASE1(BM_pop_front_concurrent_dequeue);
 BENCHMARK_CASE1(BM_pop_front_concurrent_queue);
-BENCHMARK_CASE1(BM_pop_front_concurrent_queue_sc);
 
 BENCHMARK_CASE1(BM_push_pop_par_small_mtx_queue);
 BENCHMARK_CASE1(BM_push_pop_par_small_concurrent_dequeue);
 BENCHMARK_CASE1(BM_push_pop_par_small_concurrent_queue);
-BENCHMARK_CASE1(BM_push_pop_par_small_concurrent_queue_sc);
 BENCHMARK_CASE1(BM_push_pop_par_large_mtx_queue);
 BENCHMARK_CASE1(BM_push_pop_par_large_concurrent_dequeue);
 BENCHMARK_CASE1(BM_push_pop_par_large_concurrent_queue);
-BENCHMARK_CASE1(BM_push_pop_par_large_concurrent_queue_sc);
 
 BENCHMARK_MAIN();
